requirements.txt
beautifulsoup4==4.11.2
black==23.1.0
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
google-api-core==2.11.0
google-api-python-client==2.76.0
google-auth==2.16.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.8.0
googleapis-common-protos==1.58.0
gspread==5.7.2
httplib2==0.21.0
idna==3.4
mypy-extensions==1.0.0
nameparser==1.1.2
oauthlib==3.2.2
packaging==23.0
pathspec==0.11.0
platformdirs==2.6.2
protobuf==4.21.12
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.9
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
soupsieve==2.3.2.post1
tomli==2.0.1
ujson==5.7.0
uritemplate==4.1.1
urllib3==1.26.14
